Add RollingWindow wrapper class for PyObject
