+++ Algorithm.Framework/Portfolio/BlackLittermanOptimizationPortfolioConstructionModel.py
@@ -275,7 +275,7 @@ class BlackLittermanOptimizationPortfolioConstructionModel(PortfolioConstruction
             self._symbol = symbol
             self.roc = RateOfChange(f'{symbol}.roc({lookback})', lookback)
             self.roc.updated += self.on_rate_of_change_updated
-            self.window = RollingWindow[IndicatorDataPoint](period)
+            self.window = RollingWindow(period)
 
         def reset(self):
             self.roc.updated -= self.on_rate_of_change_updated
