+++ Algorithm.Python/Alphas/ShareClassMeanReversionAlpha.py
@@ -63,7 +63,7 @@ class ShareClassMeanReversionAlphaModel(AlphaModel):
 
     def __init__(self, *args, **kwargs):
         self.sma = SimpleMovingAverage(10)
-        self.position_window = RollingWindow[float](2)
+        self.position_window = RollingWindow(2)
         self.alpha = None
         self.beta = None
         if 'tickers' not in kwargs:
